requirements.txt
PyGithub
markdown
xpinyin
